Project 5 Base code
the base code for project 5 as well as some minor updates to HW 5. Project 5 should be functioning and has been tested to ensure FET safety. Each call checks to see the previos state of the wheel and ensures that a 50ms delay is achieved before each swap occurs. This is the basis for the drive train and each wheel function has been given a argument to pass the speed.